chore: update paths to support code reorganization
