+++ test/test.options.ts
@@ -16,7 +16,7 @@ import * as assert from 'power-assert';
 
 import {Utils} from './utils';
 
-const googleapis = require('../lib/googleapis');
+const googleapis = require('../src/lib/googleapis');
 
 describe('Options', () => {
   let authClient;
