+++ test/test.path.ts
@@ -16,7 +16,7 @@ import * as assert from 'power-assert';
 
 import {Utils} from './utils';
 
-const googleapis = require('../');
+const googleapis = require('../src/lib/googleapis');
 
 describe('Path params', () => {
   let localDrive, remoteDrive;
