+++ test/test.transporters.ts
@@ -17,7 +17,7 @@ import * as assert from 'power-assert';
 
 import {Utils} from './utils';
 
-const googleapis = require('../');
+const googleapis = require('../src/lib/googleapis');
 
 function testHeaders(drive) {
   const req = drive.comments.insert(
