+++ tsconfig.tools.json
@@ -1,14 +1,13 @@
 {
-    "extends": "./node_modules/gts/tsconfig-google.json",
-    "compilerOptions": {
-      "rootDir": ".",
-      "outDir": "build",
-      "noImplicitAny": false,
-      "strictNullChecks": false,
-      "noImplicitThis": false
-    },
-    "include": [
-      "src/scripts/*.ts"
-    ]
-  }
-  
+  "extends": "./node_modules/gts/tsconfig-google.json",
+  "compilerOptions": {
+    "rootDir": ".",
+    "outDir": "build",
+    "noImplicitAny": false,
+    "strictNullChecks": false,
+    "noImplicitThis": false
+  },
+  "include": [
+    "src/scripts/*.ts"
+  ]
+}
